% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate Import ID.R
\name{generate_importID}
\alias{generate_importID}
\title{Generate Import ID}
\usage{
generate_importID(
  list_name,
  person_id = NULL,
  organization_id = NULL,
  prospect_id = NULL,
  contribution_id = NULL
)
}
\arguments{
\item{list_name}{(Character) is an single element or a vector of values to be used for creating the list name}

\item{person_id}{(Character) is a vector containing all of the primary IDs of people that should be in the list}

\item{organization_id}{(Character) is a vector containing all of the primary IDs of organizations that should be in the list}

\item{prospect_id}{(Character) is a vector containing all of the GUIDS of prospects that should be in the list}

\item{contribution_id}{(Character) is a vector containing all of the GUIDS of contributions that should be in the list}
}
\value{
returns a data frame with the rows of c(person_id, organization_id, prospect_id, and contribution_id)
}
\description{
Create an import ID list to be used for import into UI's CRM software
}
\examples{
list<-rep("My List Name", 1:20)
people<-paste0("Person", 1:10)
organizations<-paste0("Organization", 1:10)
generate_importID(list, people, organizations)
}
