% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotIn.R
\name{\%notin\%}
\alias{\%notin\%}
\title{notin}
\usage{
... \%notin\% NA
}
\description{
Search a vector for elements that are not in another vector
}
\details{
Negation of the %in% function
}
\examples{
fruits<-c("apple", "bannas", "cherries", "peaches")
favorite_fruits<-c("apple", "cherries")
disliked_fruits<-fruits[fruits \%notin\% favorite_fruits]
disliked_fruits
}
