% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Household Fix.R
\name{household_fix}
\alias{household_fix}
\title{Household Fix}
\usage{
household_fix(Missing_IDs, Fill_IDs)
}
\arguments{
\item{Missing_IDs}{(Character): The column containing NA's that need to be removed}

\item{Fill_IDs}{(Character): The matching column used to substitute the NA vales}
}
\value{
A vector of the IDs with all NA values removed
}
\description{
Combine the values of two columns of ID's into a single new column of ID's
}
\examples{
household_fix(starwars_donors$Household_Primary, starwars_donors$Primary_ID)
}
